linus_agent/pyproject.toml
[project]
name = "linus-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.11",
    "crewai[google-genai]>=1.4.1",
    "python-dotenv>=1.2.1",
]
